% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmut.data.same.R
\name{plds.data.same}
\alias{plds.data.same}
\title{Simple Check on Duplication}
\usage{
plds.data.same(DATA)
}
\arguments{
\item{DATA}{Object of class \code{data.frame} or \code{data.table}}
}
\value{
A named bool vector indicating whether the column is duplicated
}
\description{
This function checks if there is any duplicated column inside the data.
}
\examples{
plds.data.same(data.frame(ggplot2::diamonds))
}
