% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmut.auc.R
\name{pmut.auc}
\alias{pmut.auc}
\title{Calculate Model Prediction AUC with Visulization}
\usage{
pmut.auc(aa, pp, plot = FALSE)
}
\arguments{
\item{aa}{Vector of actuals, could be non-binary, but all non-zero will be treated as TRUE}

\item{pp}{Vector of predictions, could be any value, probability is most ideal}

\item{plot}{Logical (defualt is FALSE), TRUE indicates plotting the auc curve}
}
\value{
A single numeric value for auc
}
\description{
This function calculates area under the ROC curve for model prediction, without any package dependency.
}
\examples{
actuals = c(1,1,1,1,0,1,1,0,1,0,1,0,1,0,0,1,0,0,0,0)
predicts = rev(seq_along(actuals)); predicts[9:10] = mean(predicts[9:10])
pmut.auc(actuals, predicts, plot=TRUE)
}
