% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmut.base.prep.R
\name{pmut.base.prep}
\alias{pmut.base.prep}
\title{Process Data to be Socred with Meta Information}
\usage{
pmut.base.prep(DATA, CatMeta, NumMeta)
}
\arguments{
\item{DATA}{Object of class \code{data.frame} or \code{data.table}}

\item{CatMeta}{List of meta information for categorical features generated by \code{pmut.base.find}}

\item{NumMeta}{List of meta information for numeric features generated by \code{pmut.base.find}}
}
\value{
A \code{data.frame} or \code{data.table} ready to be scored
}
\description{
This function takes meta information generated by \code{\link{pmut.base.find}}, prepares new data so that it can be scored without error.
It conducts a few things: it handles missing value imputation either by assigning to base level (categorical) or mean value (numeric),
note that a new column marking the imputed numeric entry is generated in the mean time;
it assigns levels not found in meta but observed in new data to base level;
it handles levels found in meta but not observed in new data by treating the column as \code{factor};
it handles entire column found in meta but not observed in new data by imputing the entire column with its base or mean;
it attaches symbol "!" with every base level; lastly, it orders the columns alphabetically.
Note that data processed by this function will only have two classes: \code{factor} for categorical, \code{numeric} for numeric.
Then \code{model.matrix} will produce data matrix with exactly identical format as training,
so that it can be scored for a \code{glmnet} or \code{xgboost} model.
}
\examples{
temp = pmut.base.find(data.frame(ggplot2::diamonds))
# remove two columns
newdata = data.frame(ggplot2::diamonds)[,-c(2,6)]
# generate na
newdata$price[5:15] = NA
# assign new color
newdata$color = "NEW"
# temp[[1]] categorical meta, temp[[2]] numeric meta
newdata = pmut.base.prep(newdata, temp[[1]], temp[[2]])
head(newdata)
sapply(newdata, class)
}
