% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmut.data.pmis.R
\name{pmut.data.pmis}
\alias{pmut.data.pmis}
\title{Simple Check on Percentage of Missing}
\usage{
pmut.data.pmis(DATA)
}
\arguments{
\item{DATA}{Object of class \code{data.frame} or \code{data.table}}
}
\value{
A named vector having percent of missing for the column
}
\description{
This function checks percenrage of NA (include empty string for character) for every column inside the data.
}
\examples{
pmut.data.pmis(data.frame(ggplot2::diamonds))
}
