% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmut.data.scal.R
\name{pmut.data.scal}
\alias{pmut.data.scal}
\title{Simple Standardization}
\usage{
pmut.data.scal(DATA)
}
\arguments{
\item{DATA}{Object of class \code{data.frame} or \code{data.table}}
}
\value{
A \code{data.frame} or \code{data.table} after standardization
}
\description{
This function standardizes every numeric column inside the data.
}
\examples{
head(pmut.data.scal(data.frame(ggplot2::diamonds)))
}
