% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmut.edap.cont.R
\name{pmut.edap.cont}
\alias{pmut.edap.cont}
\title{Create Exploratory Visualization for Continuous Feature}
\usage{
pmut.edap.cont(datatable, varstring, targetstring, meta = c(50, 4, 0.01,
  0.99), qbin = FALSE, pred.df = NULL)
}
\arguments{
\item{datatable}{Object of class \code{data.frame} or \code{data.table}}

\item{varstring}{Single character string indicating the column name inside \code{datatable} for the discrete feature}

\item{targetstring}{Single character string indicating the column name inside \code{datatable} for the response}

\item{meta}{Numeric vector with length of 4 (default is c(50,4,0.01,0.99)): 1st indicates number of bins,
2nd indicates bin rounding digits, 3rd and 4th indicate the outlier percentile}

\item{qbin}{Logical (default is FALSE), FALSE indicates equal length bins, TRUE indicates equal weight bins (quantile view)}

\item{pred.df}{Object of class \code{data.frame} (optional), with column being prediction from each model}
}
\value{
A view of line plot stacked above the histogram
}
\description{
This function creates visualization with a line plot of a specified continuous feature against the response,
plus a distribution histogram for that feature.
In the line plot, the continuous feature will be cut into bins and then placed on the x-axis. The response will be the y-axis,
which will serve as Actual. Binning characteristics will be controlled by \code{meta} and \code{qbin}. NA will be formed as its own bin.
More lines of Prediction can be created by specifying a prediction \code{data.frame}.
}
\examples{
df = data.frame(ggplot2::diamonds)
pmut.edap.cont(df, "carat", "price")
pmut.edap.cont(df, "carat", "price", meta=c(12,2,0,1), qbin=TRUE)
pmut.edap.cont(df, "carat", "price", pred.df=data.frame(GLM=df$carat*7000-1000))
}
