% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmut.edap.disc.R
\name{pmut.edap.disc}
\alias{pmut.edap.disc}
\title{Create Exploratory Visualization for Discrete Feature}
\usage{
pmut.edap.disc(datatable, varstring, targetstring, pred.df = NULL)
}
\arguments{
\item{datatable}{Object of class \code{data.frame} or \code{data.table}}

\item{varstring}{Single character string indicating the column name inside \code{datatable} for the discrete feature}

\item{targetstring}{Single character string indicating the column name inside \code{datatable} for the response}

\item{pred.df}{Object of class \code{data.frame} (optional), with column being prediction from each model}
}
\value{
A view of line plot stacked above the histogram
}
\description{
This function creates visualization with a line plot of a specified discrete feature against the response,
plus a distribution histogram for that feature.
In the line plot, the discrete feature will be the x-axis and the response be the y-axis,
which will serve as Actual. NA will be formed as its own level.
More lines of Prediction can be created by specifying a prediction \code{data.frame}.
}
\examples{
df = data.frame(ggplot2::diamonds)
pmut.edap.disc(df, "color", "price")
pmut.edap.disc(df, "color", "price", pred.df=data.frame(GLM=as.numeric(df$color)*400+2600))
}
