% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmut.edap.R
\name{pmut.edap}
\alias{pmut.edap}
\title{Create Exploratory Visualization for A Vector of Features}
\usage{
pmut.edap(datatable, varvec, targetstring, meta = c(50, 4, 0.01, 0.99),
  qbin = FALSE, pred.df = NULL)
}
\arguments{
\item{datatable}{Object of class \code{data.frame} or \code{data.table}}

\item{varvec}{Vector of character indicating the column names inside \code{datatable} to productionalize the visulization}

\item{targetstring}{Single character string indicating the column name inside \code{datatable} for the response}

\item{meta}{Numeric vector with length of 4 (default is c(50,4,0.01,0.99)): 1st indicates number of bins,
2nd indicates bin rounding digits, 3rd and 4th indicate the outlier percentile}

\item{qbin}{Logical (default is FALSE), FALSE indicates equal length bins, TRUE indicates equal weight bins (quantile view)}

\item{pred.df}{Object of class \code{data.frame} (optional), with column being prediction from each model}
}
\value{
Views of line plot stacked above the histogram varied by the feature
}
\description{
This function creates visualization for a vector of features, using either \code{\link{pmut.edap.disc}} or \code{\link{pmut.edap.cont}},
depending on the feature class. Features of class \code{factor}, \code{character}, and \code{logical} will use \code{pmut.edap.disc};
Feature of class \code{numeric} will use \code{pmut.edap.cont};
Feature of class \code{integer} with unique values smaller than number of bins specified
by \code{meta} will use \code{pmut.edap.disc}, otherwise use \code{pmut.edap.cont}.
Some progression information will be printed on console.
}
\examples{
# Output the plots into a pdf file
pdf("EDA_Diamonds.pdf", width=12, height=10)
df = data.frame(ggplot2::diamonds)
pmut.edap(df, names(df)[-7], "price")
dev.off()
}
