% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmut.gini.R
\name{pmut.gini}
\alias{pmut.gini}
\title{Calculate Model Prediction Gini with Standardization}
\usage{
pmut.gini(aa, pp, print = FALSE)
}
\arguments{
\item{aa}{Vector of actuals, could be any value}

\item{pp}{Vector of predictions, could be any value}

\item{print}{Logical (defualt is FALSE), TRUE indicates printing the original gini before standardization}
}
\value{
A single numeric value for standardized gini
}
\description{
This function calculates the standardized gini coefficient for model prediction.
}
\examples{
actuals = c(1,1,1,1,0,1,1,0,1,0,1,0,1,0,0,1,0,0,0,0)
predicts = rev(seq_along(actuals)); predicts[9:10] = mean(predicts[9:10])
pmut.gini(actuals, predicts, print=TRUE)
}
